# Re-work pooling to have a fixed memory cost