Earlier versions of pooling didn't re-use allocations within the construction of a FST. It only amortised allocations across different FSTs being built. As a result of this, its memory usage grew proportional to the number of distinct states in the FST (this can be several gigabytes when constructing large FSTs)

To address this shortcoming, this PR relies on the fact that frozen builderNode(s) can be re-used once they're evicted from the `Registry`. A further nice property of this PR's implementation is it bounds memory usage to be proportional to the size of the `Registry` (plus max length of `unfinishedNodes`). This allows users of Vellum to control memory behaviour more predictably, and seems to give a nice 20% CPU savings too. 

### Some stats from a micro-benchmark I've been using: 

Before the PR (i.e. `master`) - time **42s**, max total memory: **8.6GB**
```
time ./vellum-upstream-master --expvar ':8081' set --sorted sample.input output.fst
./vellum-upstream-master --expvar ':8081' set --sorted sample.input output.fs  54.00s user 5.74s system 140% cpu 42.449 total
```

<img width="1411" alt="upstream-master" src="https://user-images.githubusercontent.com/199982/44627645-e176c280-a94e-11e8-8500-66a739db5e81.png">

After PR: time **33s**, max total memory: **23MB**
```
time ./vellum-custom-pooling --expvar ':8081' set --sorted sample.input output.fst
./vellum-custom-pooling --expvar ':8081' set --sorted sample.input output.fst  47.51s user 1.19s system 146% cpu 33.132 total
```

<img width="1382" alt="reworked-pooling" src="https://user-images.githubusercontent.com/199982/44627651-f3586580-a94e-11e8-8f7b-3d18ec4f54c7.png">