@steveyen 2018-08-27
wow -- very cool!  It's gonna take a bit to absorb what you did here, but the #'s are really exciting!

One other thing is we actually do code reviews over in another system, called Gerrit (same as what android, chrome, golang, libreoffice and others use).

It's running here... http://review.couchbase.org/ and here's the vellum drill down:  http://review.couchbase.org/#/q/project:vellum

You'd have to become a user of that (with a bit of a one time pain-in-the-ass setup steps, imho).