@prateek 2018-08-28
@steveyen cheers! I created http://review.couchbase.org/#/c/98887/ on Gerrit. 

Might help if I explain the thought process a bit - the code relies on the lifecycle below for builder nodes. 

```
+-------------------------+                            +-----------------------+
|    Unfinished Nodes     |      Transfer once         |        Registry       |
|(not frozen BuilderNodes)|-----BuilderNode is ------->| (frozen BuilderNodes) |
+-------------------------+         frozen             +-----------------------+
             ^                                                     |            
             |                                                     |            
             |                                                   Put()          
             | Get() on        +-------------------+             when           
             +-new char--------| BuilderNode Pool  |<-----------evicted         
                               +-------------------+                            
                                                                                
                                                                                
```

With this structure in place, we're guaranteed that worst case we'd allocate at most  `Size(Registry) + Max(len(InputKey))` buildeNodes; and further, we can re-use the builder nodes once they're evicted from the Registry.

Few things to note about the implementation: 
- We don't need to pool transitions and builderNodes separately, because builderNodes contain a slice of transitions. So as long as we reset the transitions slice to have zero length before re-using a builderNode, we get transitions pooling for free once we start pooling builder nodes. 
- builderNode storing `[]transition` instead of `[]*transition`. Mainly to avoid unnecessary GC pointer chasing. The transition type only has primitives so we can get away with this quite easily 
- [Intrusive](https://stackoverflow.com/questions/5004162/what-does-it-mean-for-a-data-structure-to-be-intrusive) linked list for the pool - I modified the builderNode structure to have a `next *builderNode` field, to allow the pooling of nodes in a single-y linked list. We only need to operate on the head to insert/remove which makes the operations quite cheap. I tried experimenting with a couple of other strategies (arena/slice allocating) but found this to be better performing and have the simplest code