@steveyen 2018-09-05
(thanks -- leaving a pointer to the gerrit change review request - http://review.couchbase.org/#/c/98887/)