@steveyen 2018-09-05
(oh heh -- I saw @prateek left the link already.  I must have been too awed by the ascii diagram to not have seen it on the first read!)