@abhinavdangeti 2018-09-06
Hey @prateek , Great work on this, I've merged your gerrit change (http://review.couchbase.org/#/c/98887/), so that should be available on the master branch of vellum. Feel free to close this PR. Cheers!